backend/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "book-content-ingestion"
version = "1.0.0"
description = "RAG chatbot for physical AI and humanoid robotics book content"
readme = "README.md"
authors = [{name = "Physical AI Team", email = "team@example.com"}]
license = {text = "MIT"}
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
]
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "qdrant-client>=1.8.0",
    "PyPDF2>=3.0.1",
    "markdown>=3.5.1",
    "python-multipart>=0.0.6",
    "google-generativeai>=0.4.0",
    "pydantic>=2.5.0",
    "python-dotenv>=1.0.0",
    "langchain>=0.0.353",
    "langchain-community>=0.0.38",
    "langchain-google-genai>=0.0.8",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"

[tool.coverage.run]
source = ["src"]
omit = ["*/tests/*", "*/test_*", "*/conftest.py"]

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'